Made MockSendMessage class, started MockValidLoginSuccess class but ran into issues with connect() (not sure if connected or not without returning a boolean).
Created GUI class with operational Login and InstantMessenger windows. 

TODO: 
Talk to Dr. Jump to better understand Status and IMEvent listeners in
order to send and receive messages as well as connecting to the server.

Need to create and insert logo into GUI. 
Complete control class and integrate into GUI.